[Refactoring] Move syntactic rename to its own file
